+++ ocaml-eglot-type-enclosing.el
@@ -49,6 +49,7 @@
     (define-key keymap (kbd "C-c C-t") #'ocaml-eglot-type-enclosing-increase-verbosity)
     (define-key keymap (kbd "C-<right>") #'ocaml-eglot-type-enclosing-increase-verbosity)
     (define-key keymap (kbd "C-<left>") #'ocaml-eglot-type-enclosing-decrease-verbosity)
+    (define-key keymap (kbd "C-e") #'ocaml-eglot-type-enclosing-refactor-extract)
     keymap)
   "Keymap for OCaml-eglot's type enclosing transient mode.")
 
@@ -109,6 +110,20 @@ If PREV-VERB is given, the verbosity change ensure that the type is different."
           (mod (1- ocaml-eglot-type-enclosing-offset)
                (length ocaml-eglot-type-enclosing-types)))
     (ocaml-eglot-type-enclosing--with-fixed-offset)))
+
+(defun ocaml-eglot-type-enclosing-refactor-extract ()
+  "Extract the current enclosing as a toplevel expression."
+  (interactive)
+  (ocaml-eglot-req--server-capable-or-lose :experimental :ocamllsp :handleRefactorExtract)
+  (when (and ocaml-eglot-type-enclosing-types
+             (>= (length ocaml-eglot-type-enclosing-types)
+                 ocaml-eglot-type-enclosing-offset))
+    (let* ((enclosing (aref ocaml-eglot-type-enclosing-types
+                            ocaml-eglot-type-enclosing-offset))
+           (start (cl-getf enclosing :start))
+           (end (cl-getf enclosing :end))
+           (result (ocaml-eglot-req--refactor-extract start end)))
+      (ocaml-eglot-util--perform-extraction result))))
 
 (defun ocaml-eglot-type-enclosing--type-buffer (type-expr)
   "Create buffer with content TYPE-EXPR of the enclosing type buffer."